go-exercises/exercise-ninja-level-13/handsOn-exercise-02/word/main_test.go
package word

import (
	"fmt"
	"github.com/sramek5/example/go-exercises/exercise-ninja-level-13/handsOn-exercise-02/quote"
	"testing"
)

func TestCount(t *testing.T) {
	n := Count("one two three")
	if n != 3 {
		t.Error("got", n, "want", 3)
	}
}

func TestUseCount(t *testing.T) {
	m := UseCount("one two three three")
	for k, v := range m {
		switch k {
		case "one":
			if v != 1 {
				t.Error("got", v, "want", 1)
			}
		case "two":
			if v != 1 {
				t.Error("got", v, "want", 1)
			}
		case "three":
			if v != 2 {
				t.Error("got", v, "want", 2)
			}
		}
	}
}

func ExampleCount() {
	fmt.Println(Count("one two three"))
	// Output:
	// 3
}

func BenchmarkCount(b *testing.B) {
	for i := 0; i < b.N; i++ {
		Count(quote.SunAlso)
	}
}

func BenchmarkUseCount(b *testing.B) {
	for i := 0; i < b.N; i++ {
		UseCount(quote.SunAlso)
	}
}